unduplicate code
